Fix timers clean up in .close()
Now they will always be cleaned up as soon as .close was called in an
active and fsEnabled logger.

Previously it was possible to trigger a race condition with .close() and
this.flush + this.stream.end() - where user could call .close() but the
timers wouldn't be cleaned until flush and stream.end were done which
could therefore result in .close() being called once again by the
reopenTimer.
This also removes the dependency on the fact that 'this.flush' never
returned an error in an active stream by therefore always cleaning up
timers independently of flush result. And cleans up timers that would be
left working when the underlying stream was somehow closed before
'close' was called.